+++ src/view/dto/inscripcion_con_participantes_dto.go
@@ -4,7 +4,7 @@ type InscripcionConParticipantesDTO struct {
 	ID             int64             `json:"id"`
 	FormaPago      string            `json:"forma_pago"`
 	MontoPagoCOP   int               `json:"monto_pago_cop"`
-	MontoPagoUSD   int               `json:"monto_pago_usd"`
+	MontoPagoUSD   float32           `json:"monto_pago_usd"`
 	UrlSoportePago string            `json:"url_soporte_pago"`
 	Estado         string            `json:"estado"`
 	FechaCreacion  string            `json:"fecha_creacion"`
